refact: rm ViewerPanel & api /tracks, set dir /Summary & new /Actions.
