+++ app/Unit/SidePanel.jsx
@@ -4,19 +4,16 @@ import {
   withRouter
 } from 'react-router-dom';
 import {connect} from "react-redux";
-import ViewerPanel from './ViewerPanel.jsx';
 import AuthorPanel from './Author/AuthorPanel.jsx';
 
-class UnitActionPanel extends React.Component {
+class SidePanel extends React.Component {
   constructor(props){
     super(props);
     this.state = {
 
     };
     this.style={
-      Com_UnitActionPanel_: {
 
-      }
     };
   }
 
@@ -25,17 +22,12 @@ class UnitActionPanel extends React.Component {
   }
 
   render(){
-    //let cx = cxBind.bind(styles);
     return(
-      <div
-        style={this.style.Com_UnitActionPanel_}>
+      <div>
         {
-          this.props.unitCurrent.identity=="author" ?(
-            <AuthorPanel
-              _set_Modalmode={this.props._set_Modalmode}/>
-          ):(
-            <ViewerPanel/>
-          )
+          (this.props.unitCurrent.identity=="author") &&
+          <AuthorPanel
+            _set_Modalmode={this.props._set_Modalmode}/>
         }
       </div>
     )
@@ -52,4 +44,4 @@ const mapStateToProps = (state)=>{
 export default withRouter(connect(
   mapStateToProps,
   null
-)(UnitActionPanel));
+)(SidePanel));
