+++ src/units/main.js
@@ -1,8 +1,6 @@
 const express = require('express');
 const main = express.Router();
 
-const tracksExecutive = require('./tracks.js');
-
 const srcExecutive = require('./src.js');
 const plainExecutive = require('./plain.js');
 const singleExecutive = require('./single/single.js');
@@ -11,8 +9,6 @@ main.param("id", (req, res, next, id)=>{
   req.reqUnitId = id;
   next();
 })
-
-main.use('/:id/track', tracksExecutive)
 
 main.use('/:id/src', srcExecutive)
 main.use('/:id', singleExecutive)
